requirements.txt
asgiref==3.3.2
pytz==2021.1
sqlparse==0.4.1
celery==5.0
Django==3.2.12
psycopg2-binary==2.9.1
redis==3.5.3
djangorestframework==3.12.4